+++ BTCPayServer.Plugins.LNbank/Pages/BasePageModel.cs
@@ -30,12 +30,14 @@ public abstract class BasePageModel : PageModel
 
     protected async Task<Wallet> GetWallet(string userId, string walletId)
     {
+        var isServerAdmin = User.IsInRole(Roles.ServerAdmin);
         return await WalletRepository.GetWallet(new WalletsQuery
         {
             UserId = new[] { UserId },
             WalletId = new[] { walletId },
             IncludeTransactions = true,
-            IsServerAdmin = User.IsInRole(Roles.ServerAdmin)
+            IsServerAdmin = isServerAdmin,
+            IncludeSoftDeleted = isServerAdmin
         });
     }
 }